requirements.txt

pipecat-ai[silero,deepgram,webrtc]>=0.0.90

openai>=1.50.0


fastapi>=0.115.0
uvicorn[standard]>=0.32.0


python-dotenv>=1.0.1

loguru>=0.7.2
